libtracefs: Remove instance parameter from synthetic events APIs
The synthetic events apply across all instances, but some histograms
configuration is performed in specific instance. That configuration is
implementation detail of the synthetic event and should not be part of
the API. These APIs are affected by the change:
 tracefs_synth_create()
 tracefs_synth_destroy()
 tracefs_synth_show()

Now, the top trace instance is used by default. If there is a use case
for a custom instance, a new API could be proposed in the future:
 tracefs_synth_add_instance()
that will associate the synthetic event with a custom instance, before
calling the create API.

Link: https://lore.kernel.org/linux-trace-devel/20211115104556.121359-9-tz.stoyanov@gmail.com

Suggested-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Suggested-by: Yordan Karadzhov (VMware) <y.karadz@gmail.com>

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>