libtracefs: Do not clean sqlhist man page source
The "make doc_clean" command assumes that all *.1 files are being
generated with "make doc" command and removes all of them. This
assumption is wrong, as the sources of man pages from section one
have prefix "*.txt.1" and must not be removed. A more precise filtering
is added, so only files generated during "make doc" are removed.

Link: https://lore.kernel.org/linux-trace-devel/20211115104556.121359-13-tz.stoyanov@gmail.com

Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>