libtracefs/Documentation: Clean up libtracefs-hist man pages
Include the tracefs_hist_key_type enum and the structure of
tracefs_hist_axis in the man page as that is expected to be useful. It
should not be expected that people must read the actual header for this
information.

Also add spaces between some of the functions to separate out
functionality a little to make it easier to read.

Also fix the missing "pass:" before "[*]axes".

Link: https://lore.kernel.org/linux-trace-devel/20211123215958.1091307-3-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>