libtracefs: Rename tracefs_event_append/verify_filter() functions
The function names:

  tracefs_event_append_filter()
  tracefs_event_verify_filter()

Are confusing, as they do not actually affect the tep_event. What they do
is to create a filter string. Rename them to a more appropriate:

  tracefs_filter_string_append()
  tracefs_filter_string_verify()

We need to keep around the old names as they have been in a tagged
release. But move them to deprecated, and have them call the appropriate
names. Also remove the old name from the man pages.

Link: https://lore.kernel.org/linux-trace-devel/20211129123157.0ce043c3@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>