libtracefs: Quiet "Nothing to be done for" messages
When doing a make after everything is already built, we have "Nothing to
de done" messages that are rather annoying.

$ make
  DESCEND            src libtracefs.so
make[1]: Nothing to be done for 'libtracefs.so'.
  DESCEND            src libtracefs.a
make[1]: '/work/git/libtracefs.git/lib/tracefs/libtracefs.a' is up to date.

This is because the target in the src directory for these are already
done, so it reports there is nothing to be done, as that's what make does
if you specify a target (and the descent into he directory specifies the
target that is displayed).

By marking the target as .PHONY will suppress that message.

Because "libtracefs.so" depends on $(LIBTRACEFS_SHARED_SO), the latter
needs to be part of the .PHONY and not the former.

Link: https://lore.kernel.org/linux-trace-devel/20211217055041.19559-11-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>