libtracefs: Remove build_prefix file on make clean
The files "build_prefix" is used to help know if the libtracefs.pc file
needs to be updated or not. If a remote directory is used, it is built in
that directory. It should also be cleaned if a make clean is done on that
directory as well. Currently deleting that file is missing from the make
clean.

Link: https://lore.kernel.org/linux-trace-devel/20211217055041.19559-16-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>
Signed-off-by: Steven Rostedt <rostedt@goodmis.org>