libtracefs: Add tracefs_instance_set_affinity() APIs
Add the APIs:

  tracefs_instance_set_affinity()
  tracefs_instance_set_affinity_set()
  tracefs_instance_set_affinity_raw()

To easily set the CPU affinity that an instance will limit what CPUs it
will trace to. The first API uses the simple string method of:

 "1,4,6-8"

To denote CPUs 1,4,6,7,8

The _set() version uses CPU_SETS and the _raw() version just writes
directly into the tracing_cpumask file.

Link: https://lore.kernel.org/linux-trace-devel/20211217180450.12d3524b@gandalf.local.home

Signed-off-by: Steven Rostedt (VMware) <rostedt@goodmis.org>