libtracefs: Do not count CPUs beyond set size in get_affinity
If the string of CPUs has bits set that are beyond the size that the
passed in cpu_set_t size can hold, do not set them.

This relies on the _S version of the CPU_SET macros from not crashing if
the size if too big. One would think that the _S versions would be made
specifically to protect against being too big.

If a CPU is set in the string but is outside the size limit of the
cpu_set_t passed in, then do not add it to the count that is returned by
tracefs_instance_get_affinity_set().

Link: https://lore.kernel.org/linux-trace-devel/20220119155741.1766541-5-rostedt@goodmis.org
Link: https://lore.kernel.org/all/CAPpZLN4n=L-ZHCXM+LDRiQu0XwR4iCnGeCKJOuOWenkz2EhESA@mail.gmail.com/

Reported-by: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>