libtracefs: Check README to know if we should do old onmatch format
Older kernels use:

  onmatch(..).synth_event(...)

Where the "synth_event" is the synthetic event that was created. The new
format uses:

  onmatch(..).trace(synth_event,...)

Where it uses the new "trace" command and the synthetic event is the first
parameter.

The tracefs README can determine which format is used. Search the README
and if the string "trace(<synthetic_event>,param list)" is found, then we
know that the kernel uses the new format. Yes, the README in the tracefs
file system *is* an ABI.

Link: https://lore.kernel.org/linux-trace-devel/20220131163642.2754485-4-rostedt@goodmis.org

Fixes: 7c9000c7495b9 ("libtracefs: Create a way to create a synthetic event")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>