libtracefs: Have synthetic events use types pid_t, gfp_t and bool
The synthetic events understand pid_t, gfp_t and bool, and it is better to
use them, as some kernels have bugs where a field with one of the above
types may not match the raw format they are.

That is, pid_t may not match s32, even though they are the same type!

Link: https://lore.kernel.org/linux-trace-devel/20220131163642.2754485-5-rostedt@goodmis.org

Fixes: 7c9000c7495b9 ("libtracefs: Create a way to create a synthetic event")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>