libtracefs: Fix up man pages to display all functions
There are several functions in the man pages that are missing from being
accessed in the man pages because, although they are described in a man
page, they are not listed at the top of the man page, which means they do
not get their own man page.

Have all functions described in the man pages get their own reference, so
that people can find them easier. And one man page listed functions that
were described in another man page. Fix that too.

Link: https://lore.kernel.org/linux-trace-devel/20220206131518.17d54368@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>