libtracefs: Add define for ALL dynamic events
Some of the tracefs APIs for dynamic events accept bitmask of events. If
a caller wants to address all types of events, it should list them in
the bitmask. This logic is fragile. When a new type of dynamic event is
introduced, the API callers should be modified. That's why a new define
is added, part of the library API, to address all types of dynamic
events.

Link: https://lore.kernel.org/linux-trace-devel/20220214080241.634290-1-tz.stoyanov@gmail.com

Suggested-by: Steven Rostedt <rostedt@goodmis.org>
Signed-off-by: Tzvetomir Stoyanov (VMware) <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>