libtracefs: Fix the ftrace filter function man pages
The ftrace filter function man page had a title for each function at the
beginning. The beginning of the file is to list the functions that are
explained in the man page with a single description of the man page.
Having one for each of the three functions breaks the creation of the man
page and two of the three functions did not have their man pages created
for them.

Link: https://lore.kernel.org/linux-trace-devel/20220222215229.3805b46e@rorschach.local.home

Fixes: 7c9565e6ba2d7 ("libtracefs: Add tracefs_function_notrace() API")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>