libtracefs: Do not initialize with NULL offsets
It appears that some compilers or static analyzers do not like
dereferencing NULL for address calculation (not sure how it they will work
with the offset_of() macro, that depends on that).

To prevent these warnings, just add a wrapper function to do the NULL
check, and then call the real function that does the initialization of the
offset, and this will make these silly sanitizers happy.

Link: https://lore.kernel.org/linux-trace-devel/20220224234823.1190425-3-irogers@google.com

Reported-by: Ian Rogers <irogers@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>