libtracefs: Add tracefs_debug_dir()
Add a function that will return where the debugfs file system is mounted
and if it is not mounted it will mount it. This is useful for trace-cmd
as it needs to read the KVM debug offset and multiplier of KVM guests,
and since tracefs_tracing_dir() does the same work to find the tracefs
file system (and may even find where debugfs is), this uses the same
code to accomplish finding the debugfs file system.

Link: https://lore.kernel.org/linux-trace-devel/20220520181412.7f0b6179@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>