libtracefs: Fix make sqlhist when built again
"make sqlhist" currently only works after a "make clean". That is because
the sqlhist.c is extracted from the man page and left there after the
first build. On the second build, the sqlhist.c still exists and the
target for "sqlhist" runs the default build as it has nothing to build
with, which fails to build as the default does not work with it.

Add the sqlhist target to .PHONY to tell make that it is not a real target
and it should not try to build it using the default methods.

Link: https://lore.kernel.org/linux-trace-devel/20220606192953.334315-2-rostedt@goodmis.org

Tested-by: Harald Seiler <hws@denx.de>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>