libtracefs: Use unique names for sql field variables
When creating two different synthetic events using tracefs_sql(), if
they share the same label, the second one will fail because
tracefs_sql() will use that label as a variable name. In the kernel,
each variable must be unique, and the second label will cause a conflict
with the variable named after the first label.

Use a unique identifier when creating the variables based on the label,
but still keep the label text in the variable name to let users still
know what label the variable is for.

Link: https://lore.kernel.org/linux-trace-devel/20220606192953.334315-5-rostedt@goodmis.org

Tested-by: Harald Seiler <hws@denx.de>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>