libtracefs: Differentiate WHERE clause when FROM and TO events are the same
With "FROM sched_switch as start JOIN sched_switch AS end" and then
filtering with "WHERE start.prev_state = 2 && end.next_prio < 100" the
parser gets confused and uses the "start.next_prio" instead of the
"end.next_prio", making the filter only on the start event and not on the
end event as expected.

Fix it by testing the field type (FIELD_TO/FIELD_FROM) as well as the
system and event of the field when selecting which event the field is a
part of.

Link: https://lore.kernel.org/linux-trace-devel/20220606192953.334315-6-rostedt@goodmis.org

Tested-by: Harald Seiler <hws@denx.de>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>