libtracefs: Set visibility of parser symbols as 'internal'
Two declarations of an object with 'hidden' linkage refer to
the same object if they are in the same shared object.

Internal visibility is like hidden visibility, but with
additional processor specific semantics. Unless otherwise specified by
the psABI, GCC defines internal visibility to mean that a function is
never called from another module.

It was observed that an 32bit linking against libtracefs failed
because the parser symbols were missing. As the parser symbols are
supposed to be used only inside the library, mark them as internal.

Link: https://lore.kernel.org/linux-trace-devel/20220607091215.24447-1-wagi@monom.org

Signed-off-by: Daniel Wagner <dwagner@suse.de>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>