libtracefs: Set the long size of the tep handle in tracefs_local_events_system()
While writing a workshop on using libtracefs, I found that it was
embarrassing that tracefs_local_events() does not set it, and failed some
parsing that relies on it. (tracefs_local_events() simply calls
tracefs_local_events_system() where the fix needs to be added).

Set the long size of the event using:

  tep_set_long_size(tep, tep_get_header_page_size(tep))

(oh, and that tep_get_header_page_size() is HORRIBLY misnamed!!!)

Link: https://lore.kernel.org/linux-trace-devel/20220722142803.24919c8a@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>