libtracefs: Add buckets field to tracefs_sql() function
Keys can be broken up into "buckets" with the "<key>.buckets=<size>"
option. As this type requires a counter, some internal functions needed to
have this counter passed around.

The external function tracefs_hist_add_key() does not suffice for the
buckets, so a new external function was added to do so:

  tracefs_hist_add_key_cnt()

that allows a counter to be passed along with the key type.

Link: https://lore.kernel.org/linux-trace-devel/20220928120939.564619a5@gandalf.local.home

Bugzilla: https://bugzilla.kernel.org/show_bug.cgi?id=216439
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>