libtracefs: Add tracefs_tracing_dir_is_mounted() API
Add the tracefs_tracing_dir_is_mounted() API that passes in a mount flag
and an address to a char pointer (or NULL). It returns 1 if the
tracing_dir was already mounted and 0 if not. If the mount flag is set,
then it will try to mount it and return -1 if it fails. If path points to
the address of a char*, then it will be set to the path that was mounted,
or was already mounted.

This is useful for wanting to unmount the tracing_dir if the tool was
the one that mounted it.

Link: https://lore.kernel.org/linux-trace-devel/20221114204522.2433500-5-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>