libtracefs: Have tracefs_tracer_set() return zero on success
The tracefs_tracer_set() does a ret = write() to write the tracer name
into the current_tracer file, but then it returns the ret as well. "write"
returns the number of bytes written, but tracefs_tracer_set() is suppose
to return 0 on success.

Have it return 0 instead of the number of bytes written on success.

Link: https://lore.kernel.org/linux-trace-devel/20221115000031.5f2da5dd@gandalf.local.home

Fixes: df130d4915b15 ("libtracefs: Add support for setting tracers")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>