libtracefs: Make the unit test depend on the static library
Several times I've been confused that the unit test was not working after
I had fixed some of the library code. The reason is that the unit test
binary does not have a dependency on the static library. Even though the
library is updated, the unit test will not reflect the changes because it
was not rebuilt against the updated static library.

Add a dependency to the static library to the unit test binary.

Link: https://lore.kernel.org/linux-trace-devel/20221115151200.7c60d0e3@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>