libtracefs: Move follow events test after iterate raw events test
The iterate raw events test tests error paths. On is passing in no
callbacks. But the code was changed where if there are follow events
attached to an instance that it is OK not to have a callback attached.
This means that the test of the follow events must be after the iterate
test, otherwise passing in a NULL callback will succeed, making the test
fail.

Move the follow events test after the iterate test.

Link: https://lore.kernel.org/linux-trace-devel/20221210201714.59fa518f@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>