libtracefs: Remove unneeded check of !dynevent in dynevent_info()
The compiler was complaining about undefined behavior because
dynevent_info() was accessing the address of &dynevent->system and later
doing a if (!dynevent). But this was fixed by creating a wrapper function
that does the checks and then passes the information to this function.

But that commit forgot to remove the check of !dynevent leaving the
compiler to believe that it can still be NULL and that the undefined
behavior of &dynevent->system still exists.

Remove the if statement and make everyone happy.

Link: https://lore.kernel.org/linux-trace-devel/20221215140203.103faf50@gandalf.local.home/
Link: https://lore.kernel.org/linux-trace-devel/20221215155534.1616b67a@gandalf.local.home

Reported-by: Ian Rogers <irogers@google.com>
Fixes: aff006d4af0c7 ("libtracefs: Do not initialize with NULL offsets")
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>