libtracefs: Do not use hwlat tracer and fdb_delete event for tests
The tracefs_reset() test required the hwlat tracer and the fdb_delete event to
exist on the system to succeed. But that trace and event are uncommon, and will
cause a failure when they are not part of the system. Instead, pick the first
tracer returned by tracefs_tracers() and also the "sched_switch" event as that
will make the test more robust with different kernel configs.

Link: https://lore.kernel.org/linux-trace-devel/20230530001951.2070c879@rorschach.local.home

Cc: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>