libtracefs: Add tracefs_instance_tracers() API
Currently there exists a tracefs_tracers() API that takes a tracing_dir
variable to define where the tracing directory is (NULL to look it up) but it
does not handle instances. As instances can have different tracers than the top
level, there needs to be an interface that shows the tracers that are available
for instances.

Add tracefs_instance_tracers() to show the available tracers for an instance
(and the available tracers for the top level if NULL is passed in. Basically
the same as tracefs_instances(NULL)).

Link: https://lore.kernel.org/linux-trace-devel/20230530002140.760b8cc5@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>