libtracefs: Clarify the tracefs_synth_create() man page
The man page for tracefs_synth_create() and tracefs_synth_set_instance()
suggests that if a synthetic event is created in a specific instance, then
it can only be used by that instance. This is incorrect. Synthetic events
created anywhere are global, and can be used by all instances. Update the
man pages to explicitly express this.

Link: https://lore.kernel.org/linux-trace-devel/20230530020232.5ad0376c@rorschach.local.home

Cc: Tzvetomir Stoyanov <tz.stoyanov@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>