libtracefs/utest: Fix crashing of synth test when synths exist
If a synthetic event already exists when running the synthetic event test,
it does a SEGFAULT. This is because the sevents passed into
test_synth_compare() is dependent on the devents size. If there's
already a synthetic event, devents will be bigger than sevents, and the
sevents will read out of bounds. Check for devents to be the same size as
sevents before calling test_synth_compare()

Link: https://lore.kernel.org/linux-trace-devel/20230530025627.19df3714@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>