libtracefs: Add tracefs_find_cid_pid() API
Add tracefs_find_cid_pid() to make it easier to find the process ID of
the main thread on the host that represents a guest. This can be used to
find the directory of /sys/kernel/debug/kvm/<PID>/ that holds
information of the given guest.

The above function will create an instance to work in. If that is
undesired and control of what tracing instance should be used,
then use tracefs_instance_find_cid_pid() to specify the instance to do
the work in.

Link: https://lore.kernel.org/linux-trace-devel/20230603161909.411c60f2@rorschach.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>