libtracefs: Add tracefs_kprobe_destory() API
Currently the only way to remove kprobes from the system via tracefs that
was created by tracefs_kprobe_raw() or tracefs_ketprobe_all is with the
big hammer of tracefs_dynevent_destroy_all(), as removing a single kprobe
requires having the tracefs_dynevent descriptor for it (which the raw
functions do not return).

Add tracefs_kprobe_destroy() that will remove an individual kprobe or
kretprobe that was created by one of the raw functions.

Link: https://lore.kernel.org/linux-trace-devel/20230605214214.2289969d@gandalf.local.home

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>