libtracefs doc: State that tracefs_dynevent_create() is needed for tracefs_kprobe_alloc()
The man page for tracefs_kprobe_alloc() states that it does not create a
kprobe in the system, but fails to state how one should be created. Add a
comment that informs the reader that the returned descriptor can be passed
to tracefs_dynevent_create() to create it in the system.

Link: https://lore.kernel.org/linux-trace-devel/20230607142800.61e56f3b@gandalf.local.home

Reported-by: Stevie Alvarez <steviea@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>