libtracefs: fix cscope makefile rule
The current 'make cscope' rule does not actually pass the filenames
gathered by 'find' to cscope.  cscope expects them either to be passed
on the command line, as this patch does with xargs, or via a
'cscope.files' intermediate file.   It doesn't read them from stdin.

As long as the file count is low, as it is in this library, the command
line version works fine.  Larger projects like the kernel need to use
the intermediate file.

Link: https://lore.kernel.org/linux-trace-devel/20230801172857.702417-1-zwisler@kernel.org

Signed-off-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>