libtracefs: Add API to remove followers from an instance or toplevel
Add the APIs:

 tracefs_follow_event_clear()
 tracefs_follow_missed_events_clear()

When adding some more tracefs unit tests, the follower callbacks were being
called unexpectedly and causing failures and crashes, as the data used by
the callbacks was already freed by the time it hit the other test.

Add a way to remove the followers and do so in the unit tests.

Link: https://lore.kernel.org/linux-trace-devel/20231228215433.54854-9-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>