libtracefs: Add API to read tracefs_cpu and return a kbuffer
Instead of having the caller have to allocate a buffer and a kbuffer to read
the content of a tracefs_cpu, add an interface to allow the tracefs_cpu to
handle this for the application.

The following three functions are added to the API:

    tracefs_cpu_read_buf()
    tracefs_cpu_buffered_read_buf()
    tracefs_cpu_flush_buf()

That act the same as the corresponding functions with the same name minus the
ending "_buf", but instead of requiring to send a buffer in and then add it to
a kbuffer element to read, just return a kbuffer structure that can be used to
read.

Link: https://lore.kernel.org/linux-trace-devel/20231228215433.54854-12-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>