libtracefs: Add TIMESTAMP_USECS_DELTA to simplify SQL timestamp compares
Since it is so common to have in the SQL synthetic event creation logic:

 (end.TIMESTAMP_USECS - start.TIMESTAMP_USECS) as delta

for creating latency synthetic events, add a shortcut that does the above.

Now the user only needs to write:

 TIMESTAMP_USECS_DELTA as delta

and it will produce the same compare.

Note, TIMESTAMP_DELTA is also equivalent to: (end.TIMESTAMP - start.TIMESTAMP)

Link: https://lore.kernel.org/linux-trace-devel/20231228215433.54854-20-rostedt@goodmis.org

Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>